Create two separate types for self-signed and common certificates.
