+++ src/main/java/net/ripe/rpki/commons/crypto/x509cert/RpkiSelfSignedCertificateBuilder.java
@@ -29,32 +29,8 @@
  */
 package net.ripe.rpki.commons.crypto.x509cert;
 
-import org.bouncycastle.asn1.x509.KeyUsage;
-
-import java.net.URI;
-
 /**
- * Builder for Embedded EE Certificates used in RpkiSignedObjects
+ * Builder for any self-signed certificates.
  */
-public class RpkiSignedObjectEeCertificateBuilder extends RpkiAuthoritySignedCertificateBuilder {
-
-    private URI cmsPublicationUri;
-
-    public void withCorrespondingCmsPublicationPoint(URI cmsPublicationUri) {
-        this.cmsPublicationUri = cmsPublicationUri;
-    }
-
-    public X509ResourceCertificate build() {
-        validateFields();
-
-        X509ResourceCertificateBuilder builder = createGenericRpkiCertificateBuilder(KeyUsage.digitalSignature);
-
-        // Implicit by standard:
-        builder.withCa(false);
-
-        builder.withSubjectInformationAccess(new X509CertificateInformationAccessDescriptor(
-            X509CertificateInformationAccessDescriptor.ID_AD_SIGNED_OBJECT, cmsPublicationUri));
-
-        return builder.build();
-    }
+public class RpkiSelfSignedCertificateBuilder extends GenericRpkiCertificateBuilder {
 }
